Graphql backend with pagination

+++ src/main/java/com/concon/simpleprojekt/source/Mutation.java
@@ -7,7 +7,8 @@ import com.coxautodev.graphql.tools.GraphQLMutationResolver;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.sql.Date;
+import java.util.Date;
+
 
 public class Mutation implements GraphQLMutationResolver {
     private PersonRepository personRepository;
@@ -28,15 +29,19 @@ public class Mutation implements GraphQLMutationResolver {
         return person;
     }
 
-    public boolean deletePerson(Person person) {
+    public boolean deletePerson(Integer id) {
+        Person person = personRepository.findById(id).get();
+        if(person == null) {
+            throw new RecordNotFoundException("The Person to be updated was not found", id);
+        }
         personRepository.delete(person);
         return true;
     }
 
-    public Person updatePerson(Person updateperson) {
-        Person person = personRepository.findById(updateperson.getId()).get();
-        if(updateperson == null) {
-            throw new RecordNotFoundException("The book to be updated was not found", updateperson.getId());
+    public Person updatePerson(Integer id,Person updateperson) {
+        Person person = personRepository.findById(id).get();
+        if(person == null) {
+            throw new RecordNotFoundException("The Person to be updated was not found", id);
         }
         person.setVorname(updateperson.getVorname());
 person.setNachname(updateperson.getNachname());
